Tests/LibWeb/Text/expected/wheel-scroll-over-iframe-falls-back-to-parent.txt

window.scrollY: 100
iframe.scrollY: 0
